[DI] Add a base class for extension with sensible defaults
